refactor:home page controller code cleanup
