+++ worklenz-backend/src/services/projects/index.ts
@@ -1,0 +1,1 @@
+export * from './project.service';
