+++ worklenz-backend/src/services/tasks/index.ts
@@ -1,0 +1,1 @@
+export * from './task.service';
